+++ PersistentStorage/Sources/Data/Repository/PersistentStorage+Codable.swift
@@ -1,22 +1,6 @@
 import Combine
 import Foundation
 
-extension UserDefaults {
-    func codableSet<Element: Codable>(_ value: Element, forKey key: String) {
-        let data = try? JSONEncoder().encode(value)
-        self.setValue(data, forKey: key)
-    }
-
-    func codableValue<Element: Codable>(forKey key: String) -> Element? {
-        guard let data = self.data(forKey: key) else {
-            return nil
-        }
-
-        let element = try? JSONDecoder().decode(Element.self, from: data)
-        return element
-    }
-}
-
 public extension PersistentStorage {
     /// Method for storing a codable value with assigned key into the user defaults storage.
     ///
